Extracts database functions to food model
